Implement ReadAt and Size for Arc<RandomAccessFile>
`Arc<RandomAccessFile>` is useful because it can be cloned, and `ReadAt`
and `Size` are not mutating operations.
`SizeCursor<Arc<RandomAccessFile>>` can be cloned and shared across
threads and also act like a file with a mutable seek position.

A more specific example is creating a
`zip::ZipArchive<SizeCursor<Arc<RandomAccessFile>>>` instance and
cloning a copy per thread.

https://docs.rs/zip/1.1.4/zip/read/struct.ZipArchive.html